requirements.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.1
aiomultiprocess==0.9.0
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blivet==3.8.2
blivet-gui==2.5.0
Brlapi==0.8.5
cachetools==5.3.2
censys==2.2.10
certifi==2023.11.17
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.2.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dataclasses-json==0.6.6
dbus-python==1.3.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
Django==5.0.6
django-haystack==3.3.0
django-modeltranslation==0.19.4
django-tinymce==2.8.0
dnf==4.19.2
dnspython==2.4.2
environ==1.0
executing==2.0.1
fastapi==0.108.0
fastjsonschema==2.19.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-speech==2.23.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==1.0.5
httpx==0.27.0
humanize==3.13.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
importlib-resources==6.1.1
initial-setup==0.3.98
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.56
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.1
libvirt-python==9.7.0
limits==3.7.0
lxml==5.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.9.0
nftables==0.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.2
openai==1.28.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.8.0
pid==2.2.3
pillow==10.2.0
platformdirs==4.2.1
ply==3.11
productmd==1.38
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycairo==1.25.1
pycares==4.4.0
pycparser==2.20
pycups==2.0.4
pycurl==7.45.2
pydantic==2.5.3
pydantic_core==2.14.6
pyee==8.2.2
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.46.0
pykickstart==3.48
pymodbus==3.6.8
pymongo==4.6.3
PyMsgBox==1.0.9
pyparsing==3.1.1
pyparted==3.13.0
pyperclip==1.8.2
pyppeteer==1.0.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-meh==0.51
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.4.16
requests==2.31.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-futures==1.0.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.1
rpm==4.19.1.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.12.0
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
Send2Trash==1.8.3
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==70.0.0
shodan==1.31.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
sos==4.6.0
soupsieve==2.5
SpeechRecognition==3.10.1
spicy==0.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.32.0.post1
statsmodels==0.14.2
stem==1.8.2
systemd-python==235
tenacity==8.3.0
terminado==0.18.1
terminator==2.1.3
theHarvester @ file:///home/jireh/theHarvester
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2==1.3.0
tldextract==5.1.1
tornado==6.4
torrequest==0.1.0
tqdm==4.66.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
xgboost==2.0.3
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
